Split lowering/imp into downsweep and lower modules
Lowering conceptually happens in two consecutive phases, but these were
defined in the same module. This commit splits the logic into
the two files lower.rs and downsweep.rs.
